requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.8.1
requests==2.27.1
redis==4.3.2